requirements.production.txt
# Production Dependencies for Fuze
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-CORS==6.0.1
gunicorn==21.2.0
psycopg2-binary==2.9.10
redis==5.0.1
python-dotenv==1.1.1
google-generativeai==0.8.3
sentence-transformers==5.0.0
torch==2.7.1
spacy==3.7.4
pgvector==0.4.1
numpy==1.26.4
scikit-learn==1.4.2
